+++ catCam_starter.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-echo "Executing CatPreyAnalyzer"
+echo "Executing Cat_Prey_Analyzer"
 # Tensorflow Stuff
-export PYTHONPATH=$PYTHONPATH:/home/carbotton/tensorflow1/models/research:/home/pi/tensorflow/models/research/slim:/home/pi/.local/lib/python3.7/site-packages
+export PYTHONPATH=$PYTHONPATH:/home/carbotton/models/research:/home/carbotton/models/research/slim:/home/carbotton/.local/lib/python3.7/site-packages
 cd /home/carbotton/smart_cat_door/Cat_Prey_Analyzer
 python3 cascade.py