+++ model_stages.py
@@ -7,11 +7,11 @@ from object_detection.utils import label_map_util
 
 cat_cam_py = str(Path(os.getcwd()).parents[0])
 print('CatCamPy:', cat_cam_py)
-PC_models_dir = os.path.join(cat_cam_py, 'CatPreyAnalyzer/models/Prey_Classifier')
-FF_models_dir = os.path.join(cat_cam_py, 'CatPreyAnalyzer/models/Face_Fur_Classifier')
-EYE_models_dir = os.path.join(cat_cam_py, 'CatPreyAnalyzer/models/Eye_Detector')
-HAAR_models_dir = os.path.join(cat_cam_py, 'CatPreyAnalyzer/models/Haar_Classifier')
-CR_models_dir = os.path.join(cat_cam_py, 'CatPreyAnalyzer/models/Cat_Recognizer')
+PC_models_dir = os.path.join(cat_cam_py, 'Cat_Prey_Analyzer/models/Prey_Classifier')
+FF_models_dir = os.path.join(cat_cam_py, 'Cat_Prey_Analyzer/models/Face_Fur_Classifier')
+EYE_models_dir = os.path.join(cat_cam_py, 'Cat_Prey_Analyzer/models/Eye_Detector')
+HAAR_models_dir = os.path.join(cat_cam_py, 'Cat_Prey_Analyzer/models/Haar_Classifier')
+CR_models_dir = os.path.join(cat_cam_py, 'Cat_Prey_Analyzer/models/Cat_Recognizer')
 
 
 class CC_MobileNet_Stage():