+++ CHANGELOG.md
@@ -10,6 +10,18 @@
   to
 	sys.path.append('/home/carbotton/smart_cat_door/Cat_Prey_Analyzer')
 	sys.path.append('/home/carbotton')  
+	
+- Change from telegram.ext import Updater, CommandHandler, Filters, MessageHandler
+  to filters with "f" instead of "F".
+  
+- Add bot credentials
+
+- Change all CatPreyAnalyzer for Cat_Prey_Analyzer
+
+- Change to local path:
+from model_stages import PC_Stage, FF_Stage, Eye_Stage, Haar_Stage, CC_MobileNet_Stage
+from camera_class import Camera
+  
 
 # catCam_starter.sh
 ### 2025.03.02
@@ -19,3 +31,8 @@
   to
 	export PYTHONPATH=$PYTHONPATH:/home/carbotton/tensorflow1/models/research:/home/pi/tensorflow/models/research/slim:/home/pi/.local/lib/python3.7/site-packages
 	cd /home/carbotton/smart_cat_door/Cat_Prey_Analyzer  
+
+
+# model_stages.py
+### 2025.03.02
+- Change all CatPreyAnalyzer to Cat_Prey_Analyzer